Most methods and calculations now use double instead of float.

+++ Vantage/Animation2D/Commands/ColorCommand.cs
@@ -1,11 +1,10 @@
 ﻿namespace Vantage.Animation2D.Commands
 {
-    using Vantage.Animation2D;
     using Vantage.Animation2D.OsbTypes;
 
     public class ColorCommand : Command<OsbColor>
     {
-        public ColorCommand(int easing, float startTime, float endTime, OsbColor startValue, OsbColor endValue)
+        public ColorCommand(int easing, double startTime, double endTime, OsbColor startValue, OsbColor endValue)
             : base("C", easing, startTime, endTime, startValue, endValue)
         {
         }
