+++ Vantage/Animation3D/Animation/EasingCurves/CubicBezierEasingCurve.cs
@@ -7,8 +7,8 @@
     public class CubicBezierEasingCurve : IEasingCurve
     {
         public static readonly CubicBezierEasingCurve Linear = new CubicBezierEasingCurve(0, 0, 0, 0);
-        public static readonly CubicBezierEasingCurve EaseIn = new CubicBezierEasingCurve(0.5f, 0, 1, 0.5f);
-        public static readonly CubicBezierEasingCurve EaseOut = new CubicBezierEasingCurve(0, 0.5f, 0.5f, 1);
+        public static readonly CubicBezierEasingCurve EaseIn = new CubicBezierEasingCurve(0.5, 0, 1, 0.5);
+        public static readonly CubicBezierEasingCurve EaseOut = new CubicBezierEasingCurve(0, 0.5, 0.5, 1);
 
         public CubicBezierEasingCurve(Vector2 p1, Vector2 p2)
         {
@@ -16,8 +16,8 @@
             this.P2 = p2;
         }
 
-        public CubicBezierEasingCurve(float x1, float y1, float x2, float y2)
-            : this(new Vector2(x1, y1), new Vector2(x2, y2))
+        public CubicBezierEasingCurve(double x1, double y1, double x2, double y2)
+            : this(new Vector2((float)x1, (float)y1), new Vector2((float)x2, (float)y2))
         {
         }
 
@@ -25,7 +25,7 @@
 
         public Vector2 P2 { get; set; }
 
-        public float Evaluate(float x)
+        public double Evaluate(double x)
         {
             var t = this.FindT(x);
             var c = 3.0f * this.P1.Y;
@@ -36,7 +36,7 @@
         }
 
         // Finds the t value that corresponds to the given x value on the curve
-        private float FindT(float x)
+        private double FindT(double x)
         {
             var x1 = this.P1.X;
             var x2 = this.P2.X;
