+++ Vantage/Animation3D/Animation/EasingCurves/LinearEasingCurve.cs
@@ -2,7 +2,7 @@ namespace Vantage.Animation3D.Animation.EasingCurves
 {
     public class LinearEasingCurve : IEasingCurve
     {
-        public float Evaluate(float x)
+        public double Evaluate(double x)
         {
             return x;
         }
