+++ Vantage/Animation3D/Animation/EasingCurves/StepEasingCurve.cs
@@ -2,7 +2,7 @@ namespace Vantage.Animation3D.Animation.EasingCurves
 {
     public class StepEasingCurve : IEasingCurve
     {
-        public float Evaluate(float x)
+        public double Evaluate(double x)
         {
             if (x >= 1)
             {
