+++ Vantage/Animation3D/Animation/Keyframes/IKeyframe.cs
@@ -2,10 +2,10 @@ namespace Vantage.Animation3D.Animation.Keyframes
 {
     public interface IKeyframe<T>
     {
-        float Time { get; set; }
+        double Time { get; set; }
 
         T Value { get; set; }
 
-        T ValueAtTime(IKeyframe<T> next, float time);
+        T ValueAtTime(IKeyframe<T> next, double time);
     }
 }
