+++ Vantage/Animation3D/Animation/Keyframes/QuaternionKeyframe.cs
@@ -6,19 +6,19 @@ namespace Vantage.Animation3D.Animation.Keyframes
 
     public class QuaternionKeyframe : Keyframe<Quaternion>
     {
-        public QuaternionKeyframe(float time, Quaternion value, IEasingCurve easingCurve)
+        public QuaternionKeyframe(double time, Quaternion value, IEasingCurve easingCurve)
             : base(time, value, easingCurve)
         {
         }
 
-        public QuaternionKeyframe(float time, Quaternion value)
+        public QuaternionKeyframe(double time, Quaternion value)
             : base(time, value)
         {
         }
 
-        internal override Quaternion Interpolate(Quaternion start, Quaternion end, float amount)
+        internal override Quaternion Interpolate(Quaternion start, Quaternion end, double amount)
         {
-            return Quaternion.Slerp(start, end, amount);
+            return Quaternion.Slerp(start, end, (float)amount);
         }
     }
 }
