+++ Vantage/Animation3D/Animation/OpacityProperty.cs
@@ -2,9 +2,9 @@ namespace Vantage.Animation3D.Animation
 {
     using Vantage.Animation3D.Animation.Keyframes;
 
-    public class OpacityProperty : AnimatableProperty<Keyframe<float>, float>
+    public class OpacityProperty : AnimatableProperty<Keyframe<double>, double>
     {
-        public OpacityProperty(float initialValue)
+        public OpacityProperty(double initialValue)
             : base(initialValue)
         {
         }
